+++ packages/column-components/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [2.0.6] - 2026-04-05 [_changes_](https://github.com/UW-Macrostrat/web-components/compare/@macrostrat/column-components-v2.0.5...@macrostrat/column-components-v2.0.6)
+
+### Patch Changes
+
+- Updated `@macrostrat/hyper` dependency
+  [6e9a6654](https://github.com/UW-Macrostrat/web-components/commit/6e9a665456460b31c30316a2a13d41abed94c43a)
+- Updated dependencies
+  [6e9a6654](https://github.com/UW-Macrostrat/web-components/commit/6e9a665456460b31c30316a2a13d41abed94c43a)
+  - @macrostrat/ui-components@5.0.7
+  - @macrostrat/timescale@3.1.2
+
 ## [2.0.5] - 2026-04-04 [_changes_](https://github.com/UW-Macrostrat/web-components/compare/@macrostrat/column-components-v2.0.4...@macrostrat/column-components-v2.0.5)
 
 ### Patch Changes