+++ packages/data-provider/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.0.2] - 2026-04-05 [_changes_](https://github.com/UW-Macrostrat/web-components/compare/@macrostrat/data-provider-v1.0.1...@macrostrat/data-provider-v1.0.2)
+
+### Patch Changes
+
+- Updated `@macrostrat/hyper` dependency
+  [6e9a6654](https://github.com/UW-Macrostrat/web-components/commit/6e9a665456460b31c30316a2a13d41abed94c43a)
+- Updated dependencies
+  [6e9a6654](https://github.com/UW-Macrostrat/web-components/commit/6e9a665456460b31c30316a2a13d41abed94c43a)
+  - @macrostrat/ui-components@5.0.7
+
 ## [1.0.1] - 2026-02-07
 
 Improve handling of absolute URLs in `fetch` wrapper.