+++ packages/map-styles/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.1.2] - 2026-04-05 [_changes_](https://github.com/UW-Macrostrat/web-components/compare/@macrostrat/map-styles-v2.1.1...@macrostrat/map-styles-v2.1.2)
+
+### Patch Changes
+
+- Updated dependencies
+  [6e9a6654](https://github.com/UW-Macrostrat/web-components/commit/6e9a665456460b31c30316a2a13d41abed94c43a)
+  - @macrostrat/ui-components@5.0.7
+
 ## [2.1.1] - 2026-04-04 [_changes_](https://github.com/UW-Macrostrat/web-components/compare/@macrostrat/map-styles-v2.1.0...@macrostrat/map-styles-v2.1.1)
 
 ### Patch Changes